+++ plug-ins/Apple.GameKit/Native/GameKitWrapper/GKMatchmaker.swift
@@ -128,7 +128,7 @@ public func GKMatchMaker_FindMatchedPlayers
     onError: @escaping NSErrorCallback
 )
 {
-    if #available(iOS 17.2, tvOS 17.2, macOS 14.2, visionOS 1.1,  *) {
+    if #available(iOS 17.2, tvOS 17.2, macOS 14.2, visionOS 1.1, *) {
         let gkMatchMaker = Unmanaged<GKMatchmaker>.fromOpaque(gkMatchmakerPtr).takeUnretainedValue()
         let gkMatchRequest = Unmanaged<GKMatchRequest>.fromOpaque(gkMatchRequestPtr).takeUnretainedValue()